modified User controller and service

+++ src/main/java/com/myclinik/service/IUserService.java
@@ -3,11 +3,10 @@ package main.java.com.myclinik.service;
 import com.myclinik.model.User;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface IUserService {
 	List<User> findAll();
-	User findOne(Long id);
+	User findOne(Integer id);
 	User createUser();
 	void saveUser(User u);
 	void deleteUser(Integer id);
